refactor: extract drawer components into separate files
