+++ src/app/dashboard/[teamId]/_components/dashboard-metric-drawer.tsx
@@ -2,48 +2,18 @@
 
 import { useEffect, useState } from "react";
 
-import type { Cadence, GoalType, MetricGoal } from "@prisma/client";
-import { formatDistanceToNow } from "date-fns";
-import {
-  BarChart3,
-  Check,
-  ClipboardCheck,
-  Loader2,
-  Pencil,
-  RefreshCw,
-  Settings,
-  Target,
-  Trash2,
-  TrendingUp,
-  Users,
-  X,
-} from "lucide-react";
+import type { Cadence } from "@prisma/client";
+import { ClipboardCheck, Loader2, X } from "lucide-react";
 import { Link } from "next-transition-router";
-import { toast } from "sonner";
 
-import { RoleAssignment } from "@/components/metric/role-assignment";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { DrawerClose } from "@/components/ui/drawer";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
 import {
   Tooltip,
   TooltipContent,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
-import { type GoalProgress, calculateTargetDisplayValue } from "@/lib/goals";
-import { formatCadence } from "@/lib/helpers/format-cadence";
-import { formatValue } from "@/lib/helpers/format-value";
-import { getDimensionDisplayLabel } from "@/lib/metrics/dimension-labels";
 import { getLatestMetricValue } from "@/lib/metrics/get-latest-value";
 import type { ChartTransformResult } from "@/lib/metrics/transformer-types";
 import { getPlatformConfig } from "@/lib/platform-config";
@@ -52,11 +22,15 @@ import { api } from "@/trpc/react";
 import type { DashboardChartWithRelations } from "@/types/dashboard";
 
 import { DashboardMetricChart } from "./dashboard-metric-chart";
+import {
+  ChartStatsBar,
+  type DrawerTab,
+  DrawerTabButtons,
+  GoalTabContent,
+  RoleTabContent,
+  SettingsTabContent,
+} from "./drawer";
 import type { PipelineStatus } from "./pipeline-status-provider";
-
-const CADENCE_OPTIONS: Cadence[] = ["DAILY", "WEEKLY", "MONTHLY"];
-
-type DrawerTab = "goal" | "role" | "settings";
 
 interface DashboardMetricDrawerProps {
   dashboardChart: DashboardChartWithRelations;
@@ -72,753 +46,6 @@ interface DashboardMetricDrawerProps {
     selectedDimension?: string,
   ) => void;
 }
-
-// =============================================================================
-// Tab Buttons Component
-// =============================================================================
-
-interface DrawerTabButtonsProps {
-  activeTab: DrawerTab;
-  onTabChange: (tab: DrawerTab) => void;
-}
-
-function DrawerTabButtons({ activeTab, onTabChange }: DrawerTabButtonsProps) {
-  const tabs: { id: DrawerTab; label: string; icon: typeof Target }[] = [
-    { id: "goal", label: "Goal", icon: Target },
-    { id: "role", label: "Role", icon: Users },
-    { id: "settings", label: "Settings", icon: Settings },
-  ];
-
-  return (
-    <div className="flex h-full flex-col">
-      {tabs.map((tab) => (
-        <button
-          key={tab.id}
-          onClick={() => onTabChange(tab.id)}
-          className={cn(
-            "flex flex-1 flex-col items-center justify-center gap-1.5",
-            "border-b transition-all duration-200 last:border-b-0",
-            activeTab === tab.id
-              ? "bg-primary/10 text-primary border-l-primary border-l-2"
-              : "text-muted-foreground hover:bg-muted/50 hover:text-foreground",
-          )}
-        >
-          <tab.icon className="h-5 w-5" />
-          <span className="text-[10px] font-medium tracking-wide uppercase">
-            {tab.label}
-          </span>
-        </button>
-      ))}
-    </div>
-  );
-}
-
-// =============================================================================
-// Chart Stats Bar Component
-// =============================================================================
-
-interface ChartStatsBarProps {
-  currentValue: { value: number; label?: string; date?: string } | null;
-  valueLabel: string | null;
-  goalProgress: GoalProgress | null;
-}
-
-function ChartStatsBar({
-  currentValue,
-  valueLabel,
-  goalProgress,
-}: ChartStatsBarProps) {
-  const timeElapsedPercent = goalProgress
-    ? (goalProgress.daysElapsed /
-        (goalProgress.daysElapsed + goalProgress.daysRemaining)) *
-      100
-    : null;
-
-  return (
-    <div className="bg-muted/30 flex items-center gap-6 border-b px-6 py-3">
-      {/* Current Value */}
-      <div className="flex items-baseline gap-2">
-        <span className="text-2xl font-bold">
-          {currentValue ? formatValue(currentValue.value) : "--"}
-        </span>
-        {(valueLabel ?? currentValue?.label) && (
-          <span className="text-muted-foreground text-sm">
-            {valueLabel ?? currentValue?.label}
-          </span>
-        )}
-      </div>
-
-      {/* Goal Progress */}
-      {goalProgress && (
-        <div className="flex items-center gap-2">
-          <Target className="text-muted-foreground h-4 w-4" />
-          <div className="bg-muted h-2 w-24 overflow-hidden rounded-full">
-            <div
-              className={cn(
-                "h-full transition-all duration-300",
-                goalProgress.progressPercent >= 100
-                  ? "bg-green-500"
-                  : goalProgress.progressPercent >= 70
-                    ? "bg-primary"
-                    : "bg-amber-500",
-              )}
-              style={{
-                width: `${Math.min(goalProgress.progressPercent, 100)}%`,
-              }}
-            />
-          </div>
-          <span className="text-sm font-medium">
-            {Math.round(goalProgress.progressPercent)}%
-          </span>
-        </div>
-      )}
-
-      {/* Time Progress */}
-      {timeElapsedPercent !== null && (
-        <div className="flex items-center gap-2">
-          <span className="text-muted-foreground text-xs">Time</span>
-          <div className="bg-muted h-2 w-24 overflow-hidden rounded-full">
-            <div
-              className="h-full bg-blue-500 transition-all duration-300"
-              style={{ width: `${Math.min(timeElapsedPercent, 100)}%` }}
-            />
-          </div>
-          <span className="text-sm font-medium">
-            {Math.round(timeElapsedPercent)}%
-          </span>
-        </div>
-      )}
-    </div>
-  );
-}
-
-// =============================================================================
-// Goal Tab Content Component
-// =============================================================================
-
-interface GoalTabContentProps {
-  metricId: string;
-  goal: MetricGoal | null;
-  goalProgress: GoalProgress | null;
-  currentValue: { value: number; label?: string; date?: string } | null;
-  valueLabel: string | null;
-  cadence: Cadence | null | undefined;
-}
-
-function GoalTabContent({
-  metricId,
-  goal,
-  goalProgress,
-  currentValue,
-  valueLabel,
-  cadence,
-}: GoalTabContentProps) {
-  const [isEditing, setIsEditing] = useState(false);
-  const [goalType, setGoalType] = useState<GoalType>(
-    goal?.goalType ?? "ABSOLUTE",
-  );
-  const [targetValue, setTargetValue] = useState(
-    goal?.targetValue?.toString() ?? "",
-  );
-
-  const utils = api.useUtils();
-
-  const upsertGoalMutation = api.goal.upsert.useMutation({
-    onSuccess: async () => {
-      toast.success("Goal saved");
-      setIsEditing(false);
-      await utils.goal.get.invalidate({ metricId });
-      await utils.dashboard.getDashboardCharts.invalidate();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-
-  const deleteGoalMutation = api.goal.delete.useMutation({
-    onSuccess: async () => {
-      toast.success("Goal deleted");
-      await utils.goal.get.invalidate({ metricId });
-      await utils.dashboard.getDashboardCharts.invalidate();
-    },
-    onError: (err) => toast.error(err.message),
-  });
-
-  const handleSaveGoal = () => {
-    const value = parseFloat(targetValue);
-    if (isNaN(value) || value <= 0) {
-      toast.error("Please enter a valid positive number");
-      return;
-    }
-    upsertGoalMutation.mutate({ metricId, goalType, targetValue: value });
-  };
-
-  const handleEditGoal = () => {
-    if (goal) {
-      setGoalType(goal.goalType);
-      setTargetValue(String(goal.targetValue));
-    }
-    setIsEditing(true);
-  };
-
-  const goalTargetValue =
-    goal && goalProgress
-      ? calculateTargetDisplayValue(
-          goal.goalType,
-          goal.targetValue,
-          goal.baselineValue ?? goalProgress.baselineValue,
-        )
-      : null;
-
-  // Editing Mode
-  if (isEditing || !goal) {
-    return (
-      <div className="flex h-full flex-col p-4">
-        <h3 className="mb-4 text-sm font-semibold">
-          {goal ? "Edit Goal" : "Set Goal"}
-        </h3>
-
-        <div className="space-y-4">
-          <div className="space-y-2">
-            <Label className="text-xs">Type</Label>
-            <ToggleGroup
-              type="single"
-              value={goalType}
-              onValueChange={(v) => v && setGoalType(v as GoalType)}
-              className="grid w-full grid-cols-2"
-            >
-              <ToggleGroupItem
-                value="ABSOLUTE"
-                className="data-[state=on]:bg-primary data-[state=on]:text-primary-foreground text-xs"
-              >
-                Absolute
-              </ToggleGroupItem>
-              <ToggleGroupItem
-                value="RELATIVE"
-                className="data-[state=on]:bg-primary data-[state=on]:text-primary-foreground text-xs"
-              >
-                Relative %
-              </ToggleGroupItem>
-            </ToggleGroup>
-          </div>
-
-          {cadence && (
-            <div className="space-y-2">
-              <Label className="text-xs">Period</Label>
-              <div className="bg-muted/50 flex h-8 items-center rounded-md border px-3">
-                <span className="text-muted-foreground text-xs capitalize">
-                  {formatCadence(cadence)}
-                </span>
-              </div>
-            </div>
-          )}
-
-          <div className="space-y-2">
-            <Label className="text-xs">
-              Target {goalType === "RELATIVE" ? "Growth %" : "Value"}
-            </Label>
-            <Input
-              type="number"
-              placeholder={
-                goalType === "RELATIVE" ? "e.g., 10 for 10%" : "Target value"
-              }
-              value={targetValue}
-              onChange={(e) => setTargetValue(e.target.value)}
-              className="h-8 text-sm"
-            />
-          </div>
-
-          <div className="flex gap-2 pt-2">
-            {goal && (
-              <Button
-                variant="outline"
-                size="sm"
-                onClick={() => setIsEditing(false)}
-                className="flex-1"
-              >
-                Cancel
-              </Button>
-            )}
-            <Button
-              size="sm"
-              onClick={handleSaveGoal}
-              disabled={upsertGoalMutation.isPending || !targetValue}
-              className="flex-1"
-            >
-              {upsertGoalMutation.isPending && (
-                <Loader2 className="mr-1 h-3 w-3 animate-spin" />
-              )}
-              Save
-            </Button>
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  // Display Mode
-  return (
-    <div className="flex h-full flex-col p-4">
-      <div className="mb-4 flex items-center justify-between">
-        <h3 className="text-sm font-semibold">Goal Progress</h3>
-        <Button
-          variant="ghost"
-          size="sm"
-          onClick={handleEditGoal}
-          className="h-7 px-2"
-        >
-          <Pencil className="mr-1 h-3 w-3" />
-          Edit
-        </Button>
-      </div>
-
-      <div className="space-y-4">
-        {/* Progress Percentage */}
-        <div className="text-center">
-          <div className="text-4xl font-bold">
-            {goalProgress
-              ? `${Math.round(goalProgress.progressPercent)}%`
-              : "--"}
-          </div>
-          <div className="text-muted-foreground text-xs">of goal achieved</div>
-        </div>
-
-        {/* Progress Bar */}
-        {goalProgress && (
-          <div className="space-y-1">
-            <div className="bg-muted h-3 w-full overflow-hidden rounded-full">
-              <div
-                className={cn(
-                  "h-full transition-all duration-300",
-                  goalProgress.progressPercent >= 100
-                    ? "bg-green-500"
-                    : goalProgress.progressPercent >= 70
-                      ? "bg-primary"
-                      : "bg-amber-500",
-                )}
-                style={{
-                  width: `${Math.min(goalProgress.progressPercent, 100)}%`,
-                }}
-              />
-            </div>
-            <div className="text-muted-foreground flex justify-between text-[10px]">
-              <span>0%</span>
-              <span>100%</span>
-            </div>
-          </div>
-        )}
-
-        {/* Time Elapsed */}
-        {goalProgress && (
-          <div className="space-y-1">
-            <Label className="text-xs">Time Elapsed</Label>
-            <div className="bg-muted h-2 w-full overflow-hidden rounded-full">
-              <div
-                className="h-full bg-blue-500 transition-all duration-300"
-                style={{
-                  width: `${Math.min(
-                    (goalProgress.daysElapsed /
-                      (goalProgress.daysElapsed + goalProgress.daysRemaining)) *
-                      100,
-                    100,
-                  )}%`,
-                }}
-              />
-            </div>
-            <div className="text-muted-foreground text-[10px]">
-              {goalProgress.daysRemaining}d remaining
-            </div>
-          </div>
-        )}
-
-        {/* Current / Target */}
-        <div className="grid grid-cols-2 gap-3 pt-2">
-          <div className="rounded-lg border p-3">
-            <div className="text-muted-foreground text-[10px] uppercase">
-              Current
-            </div>
-            <div className="text-lg font-semibold">
-              {currentValue ? formatValue(currentValue.value) : "--"}
-            </div>
-            {valueLabel && (
-              <div className="text-muted-foreground text-[10px]">
-                {valueLabel}
-              </div>
-            )}
-          </div>
-          <div className="rounded-lg border p-3">
-            <div className="text-muted-foreground text-[10px] uppercase">
-              Target
-            </div>
-            <div className="text-lg font-semibold">
-              {goal.goalType === "ABSOLUTE"
-                ? formatValue(goalTargetValue ?? goal.targetValue)
-                : `+${goal.targetValue}%`}
-            </div>
-            {cadence && (
-              <div className="text-muted-foreground text-[10px]">
-                {formatCadence(cadence)}
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* Delete Goal */}
-        <Button
-          variant="ghost"
-          size="sm"
-          onClick={() => deleteGoalMutation.mutate({ metricId })}
-          disabled={deleteGoalMutation.isPending}
-          className="text-destructive hover:text-destructive hover:bg-destructive/10 mt-4 w-full"
-        >
-          {deleteGoalMutation.isPending ? (
-            <Loader2 className="mr-1 h-3 w-3 animate-spin" />
-          ) : (
-            <Trash2 className="mr-1 h-3 w-3" />
-          )}
-          Remove Goal
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-// =============================================================================
-// Role Tab Content Component
-// =============================================================================
-
-interface RoleTabContentProps {
-  metricId: string;
-  metricName: string;
-  teamId: string | null;
-  roles: Array<{
-    id: string;
-    title: string;
-    color: string;
-    assignedUserId: string | null;
-    assignedUserName: string | null;
-  }>;
-}
-
-function RoleTabContent({
-  metricId,
-  metricName,
-  teamId,
-  roles,
-}: RoleTabContentProps) {
-  if (!teamId) {
-    return (
-      <div className="flex h-full flex-col items-center justify-center p-4">
-        <Users className="text-muted-foreground mb-2 h-8 w-8" />
-        <p className="text-muted-foreground text-center text-sm">
-          This metric is not linked to a team
-        </p>
-      </div>
-    );
-  }
-
-  return (
-    <div className="flex h-full flex-col p-4">
-      <h3 className="mb-4 text-sm font-semibold">Assigned Roles</h3>
-
-      {/* Existing role labels */}
-      {roles.length > 0 ? (
-        <div className="mb-4 flex flex-wrap gap-2">
-          {roles.map((role) => (
-            <Badge
-              key={role.id}
-              variant="outline"
-              className="flex items-center gap-1.5 py-1"
-              style={{ borderColor: role.color }}
-            >
-              <div
-                className="h-2 w-2 rounded-full"
-                style={{ backgroundColor: role.color }}
-              />
-              <span>{role.title}</span>
-              {role.assignedUserName && (
-                <span className="text-muted-foreground text-[10px]">
-                  {role.assignedUserName}
-                </span>
-              )}
-            </Badge>
-          ))}
-        </div>
-      ) : (
-        <div className="text-muted-foreground mb-4 rounded-lg border border-dashed p-4 text-center text-sm">
-          No roles assigned yet
-        </div>
-      )}
-
-      {/* Assign Role dropdown */}
-      <div className="space-y-2">
-        <Label className="text-xs">Assign Role</Label>
-        <RoleAssignment
-          metricId={metricId}
-          metricName={metricName}
-          teamId={teamId}
-          assignedRoleIds={roles.map((r) => r.id)}
-        />
-      </div>
-    </div>
-  );
-}
-
-// =============================================================================
-// Settings Tab Content Component
-// =============================================================================
-
-interface SettingsTabContentProps {
-  metricName: string;
-  metricDescription: string | null;
-  selectedChartType: string;
-  setSelectedChartType: (v: string) => void;
-  selectedCadence: Cadence;
-  setSelectedCadence: (v: Cadence) => void;
-  selectedDimension: string;
-  setSelectedDimension: (v: string) => void;
-  availableDimensions: string[] | undefined;
-  isDimensionsLoading: boolean;
-  isIntegrationMetric: boolean;
-  valueLabel: string | null;
-  hasChartChanges: boolean;
-  isProcessing: boolean;
-  isDeleting: boolean;
-  lastFetchedAt: Date | null;
-  onApplyChanges: () => void;
-  onRefresh: (forceRebuild?: boolean) => void;
-  onDelete: () => void;
-  onUpdateMetric: (name: string, description: string) => void;
-}
-
-function SettingsTabContent({
-  metricName,
-  metricDescription,
-  selectedChartType,
-  setSelectedChartType,
-  selectedCadence,
-  setSelectedCadence,
-  selectedDimension,
-  setSelectedDimension,
-  availableDimensions,
-  isDimensionsLoading,
-  isIntegrationMetric,
-  valueLabel,
-  hasChartChanges,
-  isProcessing,
-  isDeleting,
-  lastFetchedAt,
-  onApplyChanges,
-  onRefresh,
-  onDelete,
-  onUpdateMetric,
-}: SettingsTabContentProps) {
-  const [name, setName] = useState(metricName);
-  const [hasNameChanges, setHasNameChanges] = useState(false);
-
-  useEffect(() => {
-    if (!hasNameChanges) {
-      setName(metricName);
-    }
-  }, [metricName, hasNameChanges]);
-
-  const handleNameChange = (value: string) => {
-    setHasNameChanges(true);
-    setName(value);
-  };
-
-  const handleSaveName = () => {
-    if (name.trim() && name !== metricName) {
-      onUpdateMetric(name.trim(), metricDescription ?? "");
-    }
-    setHasNameChanges(false);
-  };
-
-  return (
-    <div className="flex h-full flex-col overflow-y-auto p-4">
-      <h3 className="mb-4 text-sm font-semibold">Settings</h3>
-
-      <div className="space-y-5">
-        {/* Metric Name */}
-        <div className="space-y-2">
-          <Label className="text-xs">Metric Name</Label>
-          <div className="flex gap-2">
-            <Input
-              value={name}
-              onChange={(e) => handleNameChange(e.target.value)}
-              className="h-8 text-sm"
-            />
-            <Button
-              size="icon"
-              variant="outline"
-              className="h-8 w-8 shrink-0"
-              onClick={handleSaveName}
-              disabled={!hasNameChanges || !name.trim() || name === metricName}
-            >
-              <Check className="h-3 w-3" />
-            </Button>
-          </div>
-        </div>
-
-        {/* Chart Type */}
-        <div className="space-y-2">
-          <Label className="text-xs">Chart Type</Label>
-          <ToggleGroup
-            type="single"
-            value={selectedChartType}
-            onValueChange={(v) => v && setSelectedChartType(v)}
-            className="grid w-full grid-cols-2"
-          >
-            <ToggleGroupItem
-              value="bar"
-              aria-label="Bar Chart"
-              className="data-[state=on]:bg-primary data-[state=on]:text-primary-foreground"
-            >
-              <BarChart3 className="mr-1 h-4 w-4" />
-              Bar
-            </ToggleGroupItem>
-            <ToggleGroupItem
-              value="line"
-              aria-label="Line Chart"
-              className="data-[state=on]:bg-primary data-[state=on]:text-primary-foreground"
-            >
-              <TrendingUp className="mr-1 h-4 w-4" />
-              Line
-            </ToggleGroupItem>
-          </ToggleGroup>
-        </div>
-
-        {/* Cadence */}
-        <div className="space-y-2">
-          <Label className="text-xs">Cadence</Label>
-          <ToggleGroup
-            type="single"
-            value={selectedCadence}
-            onValueChange={(v) => v && setSelectedCadence(v as Cadence)}
-            className="grid w-full grid-cols-3"
-          >
-            {CADENCE_OPTIONS.map((c) => (
-              <ToggleGroupItem
-                key={c}
-                value={c}
-                className="data-[state=on]:bg-primary data-[state=on]:text-primary-foreground text-xs"
-              >
-                {c.charAt(0) + c.slice(1).toLowerCase()}
-              </ToggleGroupItem>
-            ))}
-          </ToggleGroup>
-        </div>
-
-        {/* Dimensions */}
-        {isIntegrationMetric && (
-          <div className="space-y-2">
-            <Label className="text-xs">Dimension</Label>
-            {isDimensionsLoading ? (
-              <div className="flex h-8 items-center justify-center rounded-md border">
-                <Loader2 className="text-muted-foreground h-4 w-4 animate-spin" />
-              </div>
-            ) : availableDimensions && availableDimensions.length > 0 ? (
-              <Select
-                value={selectedDimension}
-                onValueChange={setSelectedDimension}
-              >
-                <SelectTrigger className="h-8 text-sm">
-                  <SelectValue placeholder="Select dimension" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="value">
-                    {valueLabel ?? "Primary Value"}
-                  </SelectItem>
-                  {availableDimensions.map((dim) => (
-                    <SelectItem key={dim} value={dim}>
-                      {getDimensionDisplayLabel(dim)}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            ) : (
-              <div className="text-muted-foreground h-8 rounded-md border px-3 py-1.5 text-sm">
-                No dimensions available
-              </div>
-            )}
-          </div>
-        )}
-
-        {/* Apply Changes */}
-        <Button
-          size="sm"
-          onClick={onApplyChanges}
-          disabled={isProcessing || !hasChartChanges}
-          className="w-full"
-        >
-          {isProcessing ? (
-            <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-          ) : (
-            <BarChart3 className="mr-2 h-4 w-4" />
-          )}
-          Apply Changes
-        </Button>
-
-        {/* Refresh Controls */}
-        <div className="space-y-2">
-          <Label className="text-xs">Data Refresh</Label>
-          <div className="grid grid-cols-2 gap-2">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => onRefresh(false)}
-              disabled={isProcessing}
-            >
-              <RefreshCw
-                className={cn("mr-1 h-3 w-3", isProcessing && "animate-spin")}
-              />
-              Refresh
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={() => onRefresh(true)}
-              disabled={isProcessing}
-            >
-              <RefreshCw
-                className={cn("mr-1 h-3 w-3", isProcessing && "animate-spin")}
-              />
-              Hard
-            </Button>
-          </div>
-          {lastFetchedAt && (
-            <div className="text-muted-foreground text-[10px]">
-              Last fetched{" "}
-              {formatDistanceToNow(new Date(lastFetchedAt), {
-                addSuffix: true,
-              })}
-            </div>
-          )}
-        </div>
-
-        {/* Delete Metric */}
-        <div className="border-t pt-4">
-          <Button
-            variant="destructive"
-            size="sm"
-            onClick={onDelete}
-            disabled={isDeleting}
-            className="w-full"
-          >
-            {isDeleting ? (
-              <Loader2 className="mr-2 h-4 w-4 animate-spin" />
-            ) : (
-              <Trash2 className="mr-2 h-4 w-4" />
-            )}
-            Delete Metric
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// =============================================================================
-// Main Drawer Component
-// =============================================================================
 
 export function DashboardMetricDrawer({
   dashboardChart,
