+++ src/app/dashboard/[teamId]/_components/drawer/index.ts
@@ -1,0 +1,5 @@
+export { ChartStatsBar } from "./chart-stats-bar";
+export { DrawerTabButtons, type DrawerTab } from "./drawer-tab-buttons";
+export { GoalTabContent } from "./goal-tab-content";
+export { RoleTabContent } from "./role-tab-content";
+export { SettingsTabContent } from "./settings-tab-content";
